fix(pty): split native session helpers
