+++ .omo/evidence/20260706-b1-w1-gate-fix/00-git-status-before.txt
@@ -1,0 +1,65 @@
+$ git status --short --branch
+## feat/persistent-terminal-runtime...origin/main [ahead 24]
+?? .omo/evidence/20260706-b1-audit-w0-foundations.md
+?? .omo/evidence/20260706-b1-audit-w1-native-core-retry.md
+?? .omo/evidence/20260706-b1-audit-w1-native-core.md
+?? .omo/evidence/20260706-b1-audit-w2-runtime-retry.md
+?? .omo/evidence/20260706-b1-audit-w2-runtime.md
+?? .omo/evidence/20260706-b1-audit-w6-packaging-retry.md
+?? .omo/evidence/20260706-b1-audit-w6-packaging.md
+?? .omo/evidence/20260706-b1-final-verification-qa/
+?? .omo/evidence/20260706-b1-todo1-stop-hook-2/
+?? .omo/evidence/20260706-b1-todo1-stop-hook-3/
+?? .omo/evidence/20260706-b1-todo1-stop-hook/
+?? .omo/evidence/20260706-b1-todo13-stop-hook-verify-2/
+?? .omo/evidence/20260706-b1-todo13-stop-hook-verify-3/
+?? .omo/evidence/20260706-b1-todo13-stop-hook-verify/
+?? .omo/evidence/20260706-b1-todo15-stop-hook-verify-2/
+?? .omo/evidence/20260706-b1-todo15-stop-hook-verify-3/
+?? .omo/evidence/20260706-b1-todo15-stop-hook-verify/
+?? .omo/evidence/20260706-b1-todo30-stop-hook-verify-2/
+?? .omo/evidence/20260706-b1-todo30-stop-hook-verify-3/
+?? .omo/evidence/20260706-b1-todo30-stop-hook-verify/
+?? .omo/evidence/20260706-b1-todo31-stop-hook-verify-2/
+?? .omo/evidence/20260706-b1-todo31-stop-hook-verify-3/
+?? .omo/evidence/20260706-b1-todo31-stop-hook-verify/
+?? .omo/evidence/20260706-b1-todo9-stop-hook-verify-2/
+?? .omo/evidence/20260706-b1-todo9-stop-hook-verify-3/
+?? .omo/evidence/20260706-b1-todo9-stop-hook-verify/
+?? .omo/evidence/20260706-b1-w1-gate-fix/
+?? .omo/evidence/20260706-b1-w2-runtime-stop-hook-verify-2/
+?? .omo/evidence/20260706-b1-w2-runtime-stop-hook-verify-3/
+?? .omo/evidence/20260706-b1-w2-runtime-stop-hook-verify/
+?? .omo/evidence/20260706-b1-w6-packaging-stop-hook-2/
+?? .omo/evidence/20260706-b1-w6-packaging-stop-hook-3/
+?? .omo/evidence/20260706-b1-w6-packaging-stop-hook/
+?? .omo/evidence/20260706-check-prebuild-stop-hook-verify/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-11-loader-stop-hook-2/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-11-loader-stop-hook-3/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-11-loader-stop-hook/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-12-gate-fix-stop-hook-verify-2/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-12-gate-fix-stop-hook-verify-3/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-12-gate-fix-stop-hook-verify/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-12-gate-review.md
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-12-stop-hook-verification-2/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-12-stop-hook-verification-3/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-12-stop-hook-verification/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-14-screen-model-fix/summary.md
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-14-stop-hook-verification-1/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-14-stop-hook-verification-2/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-14-stop-hook-verification-3/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-2-stop-hook-verification-2/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-2-stop-hook-verification-3/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-2-stop-hook-verification/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-3-native-workflow-stop-hook-2/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-3-native-workflow-stop-hook-3/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-3-native-workflow-stop-hook/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-4-infra-sync-stop-hook-2/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-4-infra-sync-stop-hook-3/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-4-infra-sync-stop-hook/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-6-7-stop-hook-2/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-6-7-stop-hook-3/
+?? .omo/evidence/20260706-persistent-terminal-b1/todo-6-7-stop-hook/
+?? .omo/evidence/20260706-w1-native-audit-stop-hook-verify-1/
+?? .omo/evidence/20260706-w1-native-audit-stop-hook-verify-2/
+?? .omo/evidence/20260706-w1-native-audit-stop-hook-verify-3/
