+++ .omo/evidence/20260706-b1-w1-gate-fix/01-baseline-cargo-test.txt
@@ -1,0 +1,32 @@
+$ cargo test -p senpi-pty --locked
+    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.02s
+     Running unittests src/lib.rs (target/debug/deps/senpi_pty-2d95dcab8c614e82)
+
+running 12 tests
+test tests::package_version_constant_matches_package_json ... ok
+test tests::portable_pty_backend_is_linked ... ok
+test tests::version_sentinel_matches_package_version ... ok
+test tests::version_matches_crate_version ... ok
+test session_tests::pty_session_streams_raw_bytes_and_exit_code ... ok
+test session_tests::pty_session_kill_removes_child_process_tree ... ok
+test session_tests::pty_session_sends_signal_to_process_group ... ok
+test session_tests::pty_session_write_after_exit_returns_benign_error ... ok
+test session_tests::pty_session_writes_stdin_round_trip ... ok
+test session_tests::pty_session_times_out_hung_command ... ok
+test session_tests::pty_session_resizes_reported_terminal_size ... ok
+test session_tests::pty_session_reports_nonexistent_command_error ... ok
+
+test result: ok. 12 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.61s
+
+     Running tests/manual_qa.rs (target/debug/deps/manual_qa-f86e21e3588dcbf8)
+
+running 1 test
+test manual_cat_resize_kill_transcript ... ok
+
+test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.58s
+
+   Doc-tests senpi_pty
+
+running 0 tests
+
+test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s
