+++ .omo/evidence/20260706-b1-w1-gate-fix/05-cargo-fmt-check-after-rustfmt.txt
@@ -1,0 +1,1 @@
+$ cargo fmt -p senpi-pty --check
