+++ .omo/evidence/20260706-b1-w1-gate-fix/08-cargo-clippy-senpi-pty-all-targets.txt
@@ -1,0 +1,2 @@
+$ cargo clippy -p senpi-pty --all-targets --locked -- -D warnings
+    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.05s
