+++ .omo/evidence/20260706-b1-w1-gate-fix/09-npm-test-pi-pty.txt
@@ -1,0 +1,13 @@
+$ npm test --workspace @earendil-works/pi-pty
+
+> @earendil-works/pi-pty@2026.7.5-2 test
+> vitest --run test
+
+
+ RUN  v4.1.9 /Volumes/mengmotaStorage/local-workspaces/senpi-wt/persistent-terminal-b1/packages/pty
+
+
+ Test Files  5 passed (5)
+      Tests  31 passed (31)
+   Start at  19:45:05
+   Duration  1.93s (transform 386ms, setup 0ms, import 702ms, tests 2.35s, environment 0ms)
