+++ .omo/evidence/20260706-b1-w1-gate-fix/10a-update-native-prebuild.txt
@@ -1,0 +1,3 @@
+$ node packages/pty/native/check-prebuild-fresh.mjs --update
+    Finished `release` profile [optimized] target(s) in 0.01s
+updated native prebuild for darwin-arm64: /Volumes/mengmotaStorage/local-workspaces/senpi-wt/persistent-terminal-b1/packages/pty/native/prebuilds/darwin-arm64/senpi_pty.darwin-arm64.node
